+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 - Addition: #13: Added ability to call commands on monobehaviour instances.
 - Addition: #6: Commands are now defined using a [SCCommand] attribute.
 - Addition: #18: Added comments to all existing classes.
+- Addition: #31: Added a new class of `Display`.
 
+
+- Change: #31: Refactored how the console displays are set to be visible or hidden.
+- Change: #31: Refactored some properties, fields and methods in the `SCBase` class.
 - Change: #21: Refactored how `IParsersMap` is used within the project.
 
 - Project: #9: Changed the structure of the project to allow importing into Unity as a package from a GIT repository.